refactor(client): from github folder to git
