+++ packages/amplication-client/src/Application/git/AuthAppWithGithub.scss
