+++ packages/amplication-client/src/Application/git/AuthAppWithGithubCallback.tsx
