+++ packages/amplication-client/src/Application/git/SyncWithGithubPage.tsx
