+++ packages/amplication-client/src/Application/git/dialogs/GitCreateRepo/GitCreateRepo.scss
