+++ packages/amplication-client/src/Application/git/dialogs/GitDialogsIndex.scss
