+++ packages/amplication-client/src/Application/git/dialogs/GitRepos/GitRepoItem/GitRepoItem.scss
