+++ packages/amplication-client/src/Application/git/dialogs/GitRepos/GitRepos.scss
